Runtime fixes for C# Snippet.
Changes to make the c# snippet more accessible to a wider variety of
skill level and certain issues that would occur at runtime. Added
Content-Type to header loop to prevent improper attempt to add to Header
Collection. Fixes attempt to write the content length to the request
during the write process. Fixes issue where the bytes may not be written
to underlying stream. Adds a common SSL certificate hook that can be
uncommented by developer.